Move ShadowMapping files to renderer-V2.
