+++ src/libs/snac-renderer-V2/snac-renderer-V2/RendererReimplement.h
@@ -1,6 +1,9 @@
 #pragma once
 
 
+#include <renderer/Texture.h>
+#include <renderer/ScopeGuards.h>
+
 #include <profiler/GlApi.h>
 
 #include <span>
