+++ src/libs/snac-renderer-V2/snac-renderer-V2/graph/DepthMethod.h
@@ -1,0 +1,14 @@
+#pragma once
+
+
+namespace ad::renderer {
+
+
+enum class DepthMethod
+{
+    Single,
+    Cascaded,
+};
+
+
+} // namespace ad::renderer
